+++ src/components/Logout/Logout.styled.jsx
@@ -1,0 +1,23 @@
+import styled from "@emotion/styled";
+
+export const ButtonLogout = styled.button`
+    display: flex;
+    align-items: center;
+    background-color: transparent;
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 1.37;
+    letter-spacing: 0.04em;
+    color: rgba(17, 17, 17, 0.6);
+    border: none;
+    margin-left: auto;
+    padding: 0;
+    margin-right: 12px;
+    @media (min-width: 768px) {
+        margin-left: 32px;
+    }
+    @media (min-width: 768px) {
+        margin-bottom: 18px;
+        margin-left: 15px;
+    }
+`;
