+++ src/index.css
@@ -58,15 +58,15 @@ a {
 }
 
 .loading {
-  /* position: fixed;
+  position: fixed;
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
     transform: -webkit-translate(-50%, -50%);
     transform: -moz-translate(-50%, -50%);
-    transform: -ms-translate(-50%, -50%); */
-  justify-content: center;
-  align-items: center;
+    transform: -ms-translate(-50%, -50%);
+  /* justify-content: center;
+  align-items: center; */
 }
 .icon-edit {
   width: 2em;
