+++ src/pages/User/UserPage.jsx
@@ -1,34 +1,36 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-
-// import { getUserOperation } from 'redux/User/userOperations';
 import { fetchUser } from '../../redux/User/userOperations';
 
 import UserData from '../../components/UserData/UserData';
 import PetsData from 'components/PetsData/PetsData';
-import { Container } from './UserPage.styled';
-import { getLoader } from 'redux/User/selectors';
+import { ErrorMessage } from './UserPage.styled';
+import { getLoader, getError } from 'redux/User/selectors';
+
 import { Loader } from 'components/Utils/Loader/Loader';
 
 const UserPage = () => {
     const dispatch = useDispatch();
-    const isLoading = useSelector(getLoader);
+    const Loading = useSelector(getLoader);
+    const error = useSelector(getError);
 
     useEffect(() => {
         dispatch(fetchUser());
     }, [dispatch]);
 
     return (
-        <Container>
-            {isLoading ? (
+        <>
+            {Loading ? (
                 <Loader />
+            ) : error ? (
+                <ErrorMessage> Please try again later :( </ErrorMessage>
             ) : (
                 <>
                     <UserData />
                     <PetsData />
                 </>
             )}
-        </Container>
+        </>
     );
 };
